+++ src/main/java/racingcar/service/model/Record.java
@@ -1,4 +1,4 @@
-package racingcar.service.dto;
+package racingcar.service.model;
 
 import racingcar.service.domain.Car;
 import racingcar.utils.Preconditions;
