+++ src/main/java/racingcar/ui/ResultCarOutputView.java
@@ -1,8 +1,8 @@
 package racingcar.ui;
 
 import racingcar.service.domain.Car;
-import racingcar.service.dto.Record;
-import racingcar.service.dto.RoundResult;
+import racingcar.service.model.Record;
+import racingcar.service.model.RoundResult;
 import racingcar.utils.OutputUtils;
 
 public class ResultCarOutputView implements OutputView {
